uploading typescript environment
