+++ javascript/example_code_v3/iam/README.md
@@ -1,0 +1,21 @@
+# typescript_environment_iam_examples
+Environment for AWS SDK for JavaScript (V3) IAM samples.
+
+This is a workspace where you can find working AWS SDK for JavaScript (V3) IAM samples.
+
+# Getting Started
+
+1. Install dependencies.
+```
+npm install ts-node -g
+cd typescript_environment
+yarn
+```
+
+2. In your text editor, update user variables specified in the 'Inputs' section of the sample file
+
+2. Run sample code:
+```
+cd src
+ts-node [sample name].ts // e.g., ts-node iam_accesskeylastused.ts
+```
