+++ javascript/example_code_v3/iam/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "aws-sdk-v3-iam-examples",
+  "version": "1.0.0",
+  "main": "index.js",
+  "repository": "git@github.com/awsdocs/aws-doc-sdk-examples/tree/master/javascript/example_code_v3/iam.git",
+  "author": "Brian Murray <brmur@amazon.com>, Alex Forsyth <alex-git@amazon.com>",
+  "license": "MIT",
+  "dependencies": {
+    "@aws-sdk/client-s3": "^1.0.0-gamma.6",
+    "@aws-sdk/client-iam": "^1.0.0-gamma.6",
+    "@aws-sdk/node-http-handler": "^1.0.0-gamma.5",
+    "@aws-sdk/types": "^0.1.0-gamma.6",
+    "ts-node": "^9.0.0"
+  },
+  "devDependencies": {
+    "@types/node": "^14.0.23",
+    "typescript": "^4.0.2"
+  }
+}
