+++ javascript/example_code_v3/iam/src/iam_createaccountalias.ts
@@ -5,44 +5,40 @@ ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SD
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
 scheduled for release later in 2020, and the topic containing this example will be hosted at
-https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-managing-access-keys.html.
+https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-account-aliases.html.
 
 Purpose:
-iam_listaccesskeys.js demonstrates how to retrieve information about the access keys of an IAM user.
+iam_createaccountalias.ts demonstrates how to create an alias for an AWS account.
 
 Inputs :
 - REGION
-- IAM_USER_NAME
+- ACCOUNT_ALIAS
 
 Running the code:
-node iam_listaccesskeys.js
+node iam_createaccountalias.ts
  */
-
-// snippet-start:[iam.JavaScript.keys.listAccessKeysV3]
+// snippet-start:[iam.JavaScript.alias.createAccountAliasV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const { IAMClient, ListAccessKeysCommand } = require("@aws-sdk/client-iam");
+import { IAMClient, CreateAccountAliasCommand } from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
 
 // Set the parameters
-const params = {
-  MaxItems: 5,
-  UserName: "IAM_USER_NAME", //IAM_USER_NAME
-};
+const accountAlias = { AccountAlias: "ACCOUNT_ALIAS" }; //ACCOUNT_ALIAS
 
 // Create IAM service object
 const iam = new IAMClient(REGION);
 
 const run = async () => {
   try {
-    const data = await iam.send(new ListAccessKeysCommand(params));
+    const data = await iam.send(new CreateAccountAliasCommand(accountAlias));
     console.log("Success", data);
   } catch (err) {
     console.log("Error", err);
   }
 };
 run();
-// snippet-end:[iam.JavaScript.keys.listAccessKeysV3]
+// snippet-end:[iam.JavaScript.alias.createAccountAliasV3]
 exports.run = run; //for unit tests only
