+++ javascript/example_code_v3/iam/src/iam_createuser.ts
@@ -5,43 +5,53 @@ ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SD
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
 scheduled for release later in 2020, and the topic containing this example will be hosted at
-https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-server-certificates.html.
+https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-managing-users.html.
 
 Purpose:
-iam_deleteservercert.js demonstrates how to delete an IAM SSL/TLS server certificate.
+iam_createuser.js demonstrates how to create an IAM user for an AWS account.
 
 Inputs :
 - REGION
-- CERTIFICATE_NAME
+- USER_NAME
 
 Running the code:
-node iam_deleteservercert.js
+node iam_createuser.js
  */
-// snippet-start:[iam.JavaScript.certs.deleteServerCertificateV3]
+
+// snippet-start:[iam.JavaScript.users.getUserV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const {
+import {
   IAMClient,
-  DeleteServerCertificateCommand,
-} = require("@aws-sdk/client-iam");
+  GetUserCommand,
+  CreateUserCommand,
+} from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
 
 // Set the parameters
-const params = { ServerCertificateName: "CERTIFICATE_NAME" }; // CERTIFICATE_NAME
+const params = { UserName: "USER_NAME" }; //USER_NAME
 
 // Create IAM service object
 const iam = new IAMClient(REGION);
 
 const run = async () => {
   try {
-    const data = await iam.send(new DeleteServerCertificateCommand(params));
-    console.log("Success", data);
+    const data = await iam.send(new GetUserCommand(params));
+    console.log(
+      "User " + process.argv[3] + " already exists",
+      data.User.UserId
+    );
   } catch (err) {
-    console.log("Error", err);
+    try {
+      const results = await iam.send(new CreateUserCommand(params));
+      console.log("Success", results);
+    } catch (err) {
+      console.log("Error", err);
+    }
   }
 };
 run();
-// snippet-end:[iam.JavaScript.certs.deleteServerCertificateV3]
+// snippet-end:[iam.JavaScript.users.getUserV3]
 exports.run = run; //for unit tests only
