+++ javascript/example_code_v3/iam/src/iam_deleteaccesskey.ts
@@ -5,43 +5,44 @@ ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SD
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
 scheduled for release later in 2020, and the topic containing this example will be hosted at
-https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-managing-users.html.
+https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-managing-access-keys.html.
 
 Purpose:
-iam_listusers.js demonstrates how to list IAM users.
+iam_deleteaccesskey.ts demonstrates how to delete the AWS access key pair for an IAM user.
 
-Inputs :
+Inputs(in code):
 - REGION
+- ACCESS_KEY_ID
+- USER_NAME
 
 Running the code:
-node iam_listusers.js
+  node iam_deleteaccesskey.ts
  */
-
-// snippet-start:[iam.JavaScript.users.listUsersV3]
+// snippet-start:[iam.JavaScript.keys.deleteAccessKeyV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const { IAMClient, ListUsersCommand } = require("@aws-sdk/client-iam");
+import { IAMClient, DeleteAccessKeyCommand }  from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
 
 // Set the parameters
-const params = { MaxItems: 10 };
+const params = {
+  AccessKeyId: "ACCESS_KEY_ID", // ACCESS_KEY_ID
+  UserName: "USER_NAME", // USER_NAME
+};
 
 // Create IAM service object
 const iam = new IAMClient(REGION);
 
 const run = async () => {
   try {
-    const data = await iam.send(new ListUsersCommand(params));
-    const users = data.Users || [];
-    users.forEach(function (user) {
-      console.log("User " + user.UserName + " created", user.CreateDate);
-    });
+    const data = await iam.send(new DeleteAccessKeyCommand(params));
+    console.log("Success", data);
   } catch (err) {
     console.log("Error", err);
   }
 };
 run();
-// snippet-end:[iam.JavaScript.users.listUsersV3]
+// snippet-end:[iam.JavaScript.keys.deleteAccessKeyV3]
 exports.run = run; //for unit tests only
