+++ javascript/example_code_v3/iam/src/iam_deleteaccountalias.ts
@@ -5,43 +5,43 @@ ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SD
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
 scheduled for release later in 2020, and the topic containing this example will be hosted at
-https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-managing-access-keys.html.
+https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-account-aliases.html.
 
 Purpose:
-iam_accesskeylastused.js demonstrates how to retrieve information about the last time an IAM access key was used.
+iam_deleteaccountalias.ts demonstrates how to delete an alias for an AWS account.
 
 Inputs :
 - REGION
-- ACCESS_KEY_ID
+- ALIAS
 
 Running the code:
-node iam_accesskeylastused.js
+node iam_createaccountalias.ts
  */
-// snippet-start:[iam.JavaScript.keys.getAccessKeyLastUsedV3]
+// snippet-start:[iam.JavaScript.alias.deleteAccountAliasV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const {
-  IAMClient,
-  GetAccessKeyLastUsedCommand,
-} = require("@aws-sdk/client-iam");
+import { IAMClient, DeleteAccountAliasCommand } from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
 
 // Set the parameters
-const params = { AccessKeyId: "ACCESS_KEY_ID" }; //ACCESS_KEY_ID
+const params = { AccountAlias: "ALIAS" }; // ALIAS
 
 // Create IAM service object
 const iam = new IAMClient(REGION);
 
 const run = async () => {
+  // Load the AWS SDK for Node.js
+
+  // Create IAM service object
   try {
-    const data = await iam.send(new GetAccessKeyLastUsedCommand(params));
-    console.log("Success", data.AccessKeyLastUsed);
+    const data = await iam.send(new DeleteAccountAliasCommand(params));
+    console.log("Success", data);
   } catch (err) {
     console.log("Error", err);
   }
 };
 run();
-// snippet-end:[iam.JavaScript.keys.getAccessKeyLastUsedV3]
+// snippet-end:[iam.JavaScript.alias.deleteAccountAliasV3]
 exports.run = run; //for unit tests only
