+++ javascript/example_code_v3/iam/src/iam_deleteservercert.ts
@@ -5,43 +5,43 @@ ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SD
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
 scheduled for release later in 2020, and the topic containing this example will be hosted at
-https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-account-aliases.html.
+https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-server-certificates.html.
 
 Purpose:
-iam_deleteaccountalias.js demonstrates how to delete an alias for an AWS account.
+iam_deleteservercert.ts demonstrates how to delete an IAM SSL/TLS server certificate.
 
 Inputs :
 - REGION
-- ALIAS
+- CERTIFICATE_NAME
 
 Running the code:
-node iam_createaccountalias.js
+node iam_deleteservercert.ts
  */
-// snippet-start:[iam.JavaScript.alias.deleteAccountAliasV3]
+// snippet-start:[iam.JavaScript.certs.deleteServerCertificateV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const { IAMClient, DeleteAccountAliasCommand } = require("@aws-sdk/client-iam");
+import {
+  IAMClient,
+  DeleteServerCertificateCommand,
+} from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
 
 // Set the parameters
-const params = { AccountAlias: "ALIAS" }; // ALIAS
+const params = { ServerCertificateName: "CERTIFICATE_NAME" }; // CERTIFICATE_NAME
 
 // Create IAM service object
 const iam = new IAMClient(REGION);
 
 const run = async () => {
-  // Load the AWS SDK for Node.js
-
-  // Create IAM service object
   try {
-    const data = await iam.send(new DeleteAccountAliasCommand(params));
+    const data = await iam.send(new DeleteServerCertificateCommand(params));
     console.log("Success", data);
   } catch (err) {
     console.log("Error", err);
   }
 };
 run();
-// snippet-end:[iam.JavaScript.alias.deleteAccountAliasV3]
+// snippet-end:[iam.JavaScript.certs.deleteServerCertificateV3]
 exports.run = run; //for unit tests only
