+++ javascript/example_code_v3/iam/src/iam_getpolicy.ts
@@ -5,34 +5,28 @@ ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SD
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
 scheduled for release later in 2020, and the topic containing this example will be hosted at
-https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-server-certificates.html.
+https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-policies.html.
 
 Purpose:
-iam_updateservercert.js demonstrates how to update the name of an IAM SSL/TLS server certificate.
+iam_getpolicy.ts demonstrates how to retrieve information about an IAM managed policy.
 
 Inputs :
 - REGION
-- CERTIFICATE_NAME
-- NEW_CERTIFICATE_NAME
 
 Running the code:
-node iam_updateservercert.js
+node iam_getpolicy.ts
  */
-// snippet-start:[iam.JavaScript.certs.updateServerCertificateV3]
+// snippet-start:[iam.JavaScript.policies.getPolicyV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const {
-  IAMClient,
-  UpdateServerCertificateCommand,
-} = require("@aws-sdk/client-iam");
+import { IAMClient, GetPolicyCommand } from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
 
 // Set the parameters
-var params = {
-  ServerCertificateName: "CERTIFICATE_NAME", //CERTIFICATE_NAME
-  NewServerCertificateName: "NEW_CERTIFICATE_NAME", //NEW_CERTIFICATE_NAME
+const params = {
+  PolicyArn: "arn:aws:iam::aws:policy/AWSLambdaExecute",
 };
 
 // Create IAM service object
@@ -40,12 +34,12 @@ const iam = new IAMClient(REGION);
 
 const run = async () => {
   try {
-    const data = await iam.send(new UpdateServerCertificateCommand(params));
-    console.log("Success", data);
+    const data = await iam.send(new GetPolicyCommand(params));
+    console.log("Success", data.Policy.Description);
   } catch (err) {
     console.log("Error", err);
   }
 };
 run();
-// snippet-end:[iam.JavaScript.certs.updateServerCertificateV3]
+// snippet-end:[iam.JavaScript.policies.getPolicyV3]
 exports.run = run; //for unit tests only
