+++ javascript/example_code_v3/iam/src/iam_getservercert.ts
@@ -5,41 +5,42 @@ ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SD
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
 scheduled for release later in 2020, and the topic containing this example will be hosted at
-https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-policies.html.
+https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-server-certificates.html.
 
 Purpose:
-iam_getpolicy.js demonstrates how to retrieve information about an IAM managed policy.
+iam_getservercert.ts demonstrates how to retrieve information about an IAM SSL/TLS server certificate.
 
 Inputs :
 - REGION
+- CERTIFICATE_NAME
 
 Running the code:
-node iam_getpolicy.js
+node iam_getservercert.ts
  */
-// snippet-start:[iam.JavaScript.policies.getPolicyV3]
+// snippet-start:[iam.JavaScript.certs.getServerCertificateV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const { IAMClient, GetPolicyCommand } = require("@aws-sdk/client-iam");
+import {
+  IAMClient,
+  GetServerCertificateCommand,
+} from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
 
 // Set the parameters
-const params = {
-  PolicyArn: "arn:aws:iam::aws:policy/AWSLambdaExecute",
-};
+const params = { ServerCertificateName: "CERTIFICATE_NAME" }; //CERTIFICATE_NAME
 
 // Create IAM service object
 const iam = new IAMClient(REGION);
 
 const run = async () => {
   try {
-    const data = await iam.send(new GetPolicyCommand(params));
-    console.log("Success", data.Policy.Description);
+    const data = await iam.send(new GetServerCertificateCommand(params));
+    console.log("Success", data);
   } catch (err) {
     console.log("Error", err);
   }
 };
-run();
-// snippet-end:[iam.JavaScript.policies.getPolicyV3]
+// snippet-end:[iam.JavaScript.certs.getServerCertificateV3]
 exports.run = run; //for unit tests only
