+++ javascript/example_code_v3/iam/src/iam_listaccountaliases.ts
@@ -5,42 +5,40 @@ ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SD
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
 scheduled for release later in 2020, and the topic containing this example will be hosted at
-https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-server-certificates.html.
+https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-account-aliases.html.
 
 Purpose:
-iam_getservercert.js demonstrates how to retrieve information about an IAM SSL/TLS server certificate.
+iam_listaccountaliases.ts demonstrates how to retrieve information about the aliases for an AWS account.
 
 Inputs :
 - REGION
-- CERTIFICATE_NAME
 
 Running the code:
-node iam_getservercert.js
+node iam_listaccountaliases.ts
  */
-// snippet-start:[iam.JavaScript.certs.getServerCertificateV3]
+
+// snippet-start:[iam.JavaScript.alias.listAccountAliasesV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const {
-  IAMClient,
-  GetServerCertificateCommand,
-} = require("@aws-sdk/client-iam");
+import { IAMClient, ListAccountAliasesCommand } from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
 
 // Set the parameters
-const params = { ServerCertificateName: "CERTIFICATE_NAME" }; //CERTIFICATE_NAME
+const params = { MaxItems: 5 };
 
 // Create IAM service object
 const iam = new IAMClient(REGION);
 
 const run = async () => {
   try {
-    const data = await iam.send(new GetServerCertificateCommand(params));
+    const data = await iam.send(new ListAccountAliasesCommand(params));
     console.log("Success", data);
   } catch (err) {
     console.log("Error", err);
   }
 };
-// snippet-end:[iam.JavaScript.certs.getServerCertificateV3]
+run();
+// snippet-end:[iam.JavaScript.alias.listAccountAliasesV3]
 exports.run = run; //for unit tests only
