+++ javascript/example_code_v3/iam/src/iam_listservercerts.ts
@@ -1,49 +1,43 @@
 /* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
+
 ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SDK for JavaScript,
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
 scheduled for release later in 2020, and the topic containing this example will be hosted at
-https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-managing-access-keys.html.
+https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-server-certificates.html.
 
 Purpose:
-iam_updateaccesskey.js demonstrates how to update the status of an IAM user's access key.
+iam_listservercerts.ts demonstrates how to list the IAM SSL/TLS server certificates.
 
 Inputs :
 - REGION
-- ACCESS_KEY_ID
-- USER_NAME
 
 Running the code:
-node iam_updateaccesskey.js
+node iam_listservercerts.ts
  */
-
-// snippet-start:[iam.JavaScript.keys.updateAccessKeyV3]
+// snippet-start:[iam.JavaScript.certs.listServerCertificatesV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const { IAMClient, UpdateAccessKeyCommand } = require("@aws-sdk/client-iam");
+import {
+  IAMClient,
+  ListServerCertificatesCommand,
+} from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
-
-// Set the parameters
-var params = {
-  AccessKeyId: "ACCESS_KEY_ID", //ACCESS_KEY_ID
-  Status: "Active",
-  UserName: "USER_NAME", //USER_NAME
-};
 
 // Create IAM service object
 const iam = new IAMClient(REGION);
 
 const run = async () => {
   try {
-    const data = await iam.send(new UpdateAccessKeyCommand(params));
+    const data = await iam.send(new ListServerCertificatesCommand({}));
     console.log("Success", data);
   } catch (err) {
     console.log("Error", err);
   }
 };
 run();
-// snippet-end:[iam.JavaScript.keys.updateAccessKeyV3]
+// snippet-end:[iam.JavaScript.certs.listServerCertificatesV3]
 exports.run = run; //for unit tests only
