+++ javascript/example_code_v3/iam/src/iam_listusers.ts
@@ -5,41 +5,43 @@ ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SD
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
 scheduled for release later in 2020, and the topic containing this example will be hosted at
-https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-managing-access-keys.html.
+https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-managing-users.html.
 
 Purpose:
-iam_createaccesskeys.js demonstrates how to create a new AWS access key and AWS access key ID for an IAM user.
+iam_listusers.ts demonstrates how to list IAM users.
 
 Inputs :
 - REGION
- - IAM_USER_NAME
 
 Running the code:
-node iam_createaccesskeys.js >newuserkeys.txt
-(This create newuserkeys.txt and adds the access key ID and secret key to it.)
+node iam_listusers.ts
  */
-// snippet-start:[iam.JavaScript.keys.createAccessKeyV3]
+
+// snippet-start:[iam.JavaScript.users.listUsersV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const { IAMClient, CreateAccessKeyCommand } = require("@aws-sdk/client-iam");
+import { IAMClient, ListUsersCommand } from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
 
 // Set the parameters
-const userName = "IAM_USER_NAME"; //IAM_USER_NAME
+const params = { MaxItems: 10 };
 
 // Create IAM service object
 const iam = new IAMClient(REGION);
 
 const run = async () => {
   try {
-    const data = await iam.send(new CreateAccessKeyCommand(userName));
-    console.log("Success", data.AccessKey);
+    const data = await iam.send(new ListUsersCommand(params));
+    const users = data.Users || [];
+    users.forEach(function (user) {
+      console.log("User " + user.UserName + " created", user.CreateDate);
+    });
   } catch (err) {
     console.log("Error", err);
   }
 };
 run();
-// snippet-end:[iam.JavaScript.keys.createAccessKeyV3]
+// snippet-end:[iam.JavaScript.users.listUsersV3]
 exports.run = run; //for unit tests only
