+++ javascript/example_code_v3/iam/src/iam_updateaccesskey.ts
@@ -1,6 +1,5 @@
 /* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
-
 ABOUT THIS NODE.JS EXAMPLE: This example works with Version 3 (V3) of the AWS SDK for JavaScript,
 which is scheduled for release later in 2020. The prerelease version of the SDK is available
 at https://github.com/aws/aws-sdk-js-v3. The 'SDK for JavaScript Developer Guide' for V3 is also
@@ -8,28 +7,30 @@ scheduled for release later in 2020, and the topic containing this example will 
 https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/iam-examples-managing-access-keys.html.
 
 Purpose:
-iam_deleteaccesskey.js demonstrates how to delete the AWS access key pair for an IAM user.
+iam_updateaccesskey.ts demonstrates how to update the status of an IAM user's access key.
 
-Inputs(in code):
+Inputs :
 - REGION
 - ACCESS_KEY_ID
 - USER_NAME
 
 Running the code:
-  node iam_deleteaccesskey.js
+node iam_updateaccesskey.ts
  */
-// snippet-start:[iam.JavaScript.keys.deleteAccessKeyV3]
+
+// snippet-start:[iam.JavaScript.keys.updateAccessKeyV3]
 
 // Import required AWS SDK clients and commands for Node.js
-const { IAMClient, DeleteAccessKeyCommand } = require("@aws-sdk/client-iam");
+import { IAMClient, UpdateAccessKeyCommand } from "@aws-sdk/client-iam";
 
 // Set the AWS Region
 const REGION = "REGION"; //e.g. "us-east-1"
 
 // Set the parameters
-const params = {
-  AccessKeyId: "ACCESS_KEY_ID", // ACCESS_KEY_ID
-  UserName: "USER_NAME", // USER_NAME
+var params = {
+  AccessKeyId: "ACCESS_KEY_ID", //ACCESS_KEY_ID
+  Status: "Active",
+  UserName: "USER_NAME", //USER_NAME
 };
 
 // Create IAM service object
@@ -37,12 +38,12 @@ const iam = new IAMClient(REGION);
 
 const run = async () => {
   try {
-    const data = await iam.send(new DeleteAccessKeyCommand(params));
+    const data = await iam.send(new UpdateAccessKeyCommand(params));
     console.log("Success", data);
   } catch (err) {
     console.log("Error", err);
   }
 };
 run();
-// snippet-end:[iam.JavaScript.keys.deleteAccessKeyV3]
+// snippet-end:[iam.JavaScript.keys.updateAccessKeyV3]
 exports.run = run; //for unit tests only
