+++ javascript/example_code_v3/iam/src/metadata.yaml
@@ -1,0 +1,67 @@
+---
+# AWS IAM examples
+files:
+  - path: iam_accesskeylastused.ts
+    services:
+      - iam
+  - path: iam_attachrolepolicy.ts
+    services:
+      - iam
+  - path: iam_createaccesskeys.ts
+    services:
+      - iam
+  - path: iam_createaccountalias.ts
+    services:
+      - iam
+  - path: iam_createpolicy.ts
+    services:
+      - iam
+  - path: iam_createuser.ts
+    services:
+      - iam
+  - path: iam_deleteaccesskey.ts
+    services:
+      - iam
+  - path: iam_deleteaccountalias.ts
+    services:
+      - iam
+  - path: iam_deleteservercert.ts
+    services:
+      - iam
+  - path: iam_deleteuser.ts
+    services:
+      - iam
+  - path: iam_detachrolepolicy.ts
+    services:
+      - iam
+  - path: iam_getpolicy.ts
+    services:
+      - iam
+  - path: iam_getservercert.ts
+    services:
+      - iam
+  - path: iam_listaccesskeys.ts
+    services:
+      - iam
+  - path: iam_listaccountaliases.ts
+    services:
+      - iam
+  - path: iam_listservercerts.ts
+    services:
+      - iam
+  - path: iam_listusers.ts
+    services:
+      - iam
+  - path: iam_updateaccesskey.ts
+    services:
+      - iam
+  - path: iam_updateservercert.ts
+    services:
+      - iam
+  - path: iam_updateuser.ts
+    services:
+      - iam
+  - path: sts_assumerole.ts
+    services:
+      - sts
+...
